# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server CRM server.",
        "title": "Golang CRM Swagger Documentation",
        "contact": {
            "name": "API Support",
            "url": "https://t.me/xalmatoff",
            "email": "venom.uzz@mail.ru"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/v1/orders": {
            "post": {
                "description": "This API to create an order.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Create an order.",
                "parameters": [
                    {
                        "description": "data body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateOrderWithProductsInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Orders"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
        "/v1/orders-by-filter": {
            "get": {
                "description": "This API to get orders by filter.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Get orders by filter.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Orders"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
        "/v1/orders/{id}": {
            "get": {
                "description": "This API to get order by id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Get order by id.",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "id",
                        "description": "id for get Order",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Orders"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "This API to update an order.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Update an order.",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "id",
                        "description": "id for update Order",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "data body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateOrderInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Orders"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "This API to delete an order.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Delete an order.",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "id",
                        "description": "id for delete Order",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CreateOrderProductsInput": {
            "type": "object",
            "required": [
                "amount",
                "productId"
            ],
            "properties": {
                "productId": {
                    "type": "string",
                    "x-order": "1"
                },
                "amount": {
                    "type": "number",
                    "x-order": "2"
                }
            }
        },
        "models.CreateOrderWithProductsInput": {
            "type": "object",
            "properties": {
                "system": {
                    "type": "string",
                    "x-order": "1"
                },
                "home": {
                    "type": "string",
                    "x-order": "10"
                },
                "apartment": {
                    "type": "string",
                    "x-order": "11"
                },
                "comment": {
                    "type": "string",
                    "x-order": "12"
                },
                "paymentType": {
                    "type": "string",
                    "x-order": "13"
                },
                "deliveryPrice": {
                    "type": "number",
                    "x-order": "14"
                },
                "fullSum": {
                    "type": "number",
                    "x-order": "15"
                },
                "status": {
                    "type": "integer",
                    "x-order": "16"
                },
                "orderProducts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CreateOrderProductsInput"
                    },
                    "x-order": "17"
                },
                "accountID": {
                    "type": "integer",
                    "x-order": "2"
                },
                "chatID": {
                    "type": "integer",
                    "x-order": "3"
                },
                "regionID": {
                    "type": "integer",
                    "x-order": "4"
                },
                "customerName": {
                    "type": "string",
                    "x-order": "5"
                },
                "customerPhone": {
                    "type": "string",
                    "x-order": "6"
                },
                "city": {
                    "type": "string",
                    "x-order": "7"
                },
                "district": {
                    "type": "string",
                    "x-order": "8"
                },
                "street": {
                    "type": "string",
                    "x-order": "9"
                }
            }
        },
...
        "models.Orders": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "x-order": "1"
                },
                "home": {
                    "type": "string",
                    "x-order": "10"
                },
                "apartment": {
                    "type": "string",
                    "x-order": "11"
                },
                "comment": {
                    "type": "string",
                    "x-order": "12"
                },
                "paymentType": {
                    "type": "string",
                    "x-order": "13"
                },
                "deliveryPrice": {
                    "type": "number",
                    "x-order": "14"
                },
                "fullSum": {
                    "type": "number",
                    "x-order": "15"
                },
                "status": {
                    "type": "integer",
                    "x-order": "16"
                },
                "createdAt": {
                    "type": "string",
                    "x-order": "17"
                },
                "updatedAt": {
                    "type": "string",
                    "x-order": "18"
                },
                "deletedAt": {
                    "type": "string",
                    "x-order": "19"
                },
                "system": {
                    "type": "string",
                    "x-order": "2"
                },
                "chatID": {
                    "type": "integer",
                    "x-order": "3"
                },
                "accountID": {
                    "type": "integer",
                    "x-order": "3"
                },
                "regionID": {
                    "type": "integer",
                    "x-order": "4"
                },
                "customerName": {
                    "type": "string",
                    "x-order": "5"
                },
                "customerPhone": {
                    "type": "string",
                    "x-order": "6"
                },
                "city": {
                    "type": "string",
                    "x-order": "7"
                },
                "district": {
                    "type": "string",
                    "x-order": "8"
                },
                "street": {
                    "type": "string",
                    "x-order": "9"
                }
            }
        },
...
        "models.UpdateOrderInput": {
            "type": "object",
            "properties": {
                "system": {
                    "type": "string",
                    "x-order": "1"
                },
                "home": {
                    "type": "string",
                    "x-order": "10"
                },
                "apartment": {
                    "type": "string",
                    "x-order": "11"
                },
                "comment": {
                    "type": "string",
                    "x-order": "12"
                },
                "paymentType": {
                    "type": "string",
                    "x-order": "13"
                },
                "deliveryPrice": {
                    "type": "number",
                    "x-order": "14"
                },
                "fullSum": {
                    "type": "number",
                    "x-order": "15"
                },
                "status": {
                    "type": "integer",
                    "x-order": "16"
                },
                "accountID": {
                    "type": "integer",
                    "x-order": "2"
                },
                "chatID": {
                    "type": "integer",
                    "x-order": "3"
                },
                "regionID": {
                    "type": "integer",
                    "x-order": "4"
                },
                "customerName": {
                    "type": "string",
                    "x-order": "5"
                },
                "customerPhone": {
                    "type": "string",
                    "x-order": "6"
                },
                "city": {
                    "type": "string",
                    "x-order": "7"
                },
                "district": {
                    "type": "string",
                    "x-order": "8"
                },
                "street": {
                    "type": "string",
                    "x-order": "9"
                }
            }
        },
...
        "v1.Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}